move utils.py
